+++ lib/hub.dart
@@ -7,8 +7,17 @@ export 'src/hub/hub.dart';
 export 'src/hub/hub_client.dart';
 export 'src/hub/hub_provider.dart';
 export 'src/hub/hub_resource.dart';
-export 'src/hub/resource.dart';
-export 'src/hub/resource_adapter.dart';
 
-export 'src/hub/rest/resource_rest_client.dart';
+export 'src/hub/resource.dart';
+export 'src/hub/element_resource.dart';
+export 'src/hub/collection_resource.dart';
+
+export 'src/hub/element_resource_adapter.dart';
+export 'src/hub/collection_resource_adapter.dart';
+
+export 'src/hub/rest/element_resource_rest_client.dart';
+export 'src/hub/rest/collection_resource_rest_client.dart';
+
 export 'src/hub/rest/resource_rest_endpoint.dart';
+export 'src/hub/rest/element_resource_rest_endpoint.dart';
+export 'src/hub/rest/collection_resource_rest_endpoint.dart';