+++ test/resource_hub/lib/memo.dart
@@ -4,8 +4,10 @@ part 'memo.g.dart';
 
 @TransferObject()
 class Memo extends _TransferObject {
+  @TransferId()
+  final int id;
   final String text;
   final DateTime timestamp;
 
-  Memo(this.text, this.timestamp);
+  Memo(this.id, this.text, this.timestamp);
 }